requirements.txt
# Core dependencies for Codenames AI Benchmark

# BAML - Structured LLM outputs and prompt management
baml-py>=0.211.0

# LLM providers (still needed by BAML clients)
openai>=1.0.0  # Also used for xAI Grok and DeepSeek (OpenAI-compatible APIs)
anthropic>=0.18.0
google-generativeai>=0.3.0

# Optional: for loading .env files
python-dotenv>=1.0.0

